refactoring of uuid creation func and waypoint generation func.
