+++ lib/common.inc.php
@@ -199,6 +199,24 @@ $emailheaders = "Content-Transfer-Encoding: 8bit\r\n";
 $emailheaders .= 'From: "' . $emailaddr . '" <' . $emailaddr . '>';
 
 
+/**
+ * -- This script is moved here from clicompatbase - should be removed from here in the future --
+ *
+ * Create a "universal unique" replication "identifier"
+ */
+function create_uuid()
+{
+    $uuid = mb_strtoupper(md5(uniqid(rand(), true)));
+
+    //split into XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX (type VARCHAR 36, case insensitiv)
+    $uuid = mb_substr($uuid, 0, 8) . '-' . mb_substr($uuid, -24);
+    $uuid = mb_substr($uuid, 0, 13) . '-' . mb_substr($uuid, -20);
+    $uuid = mb_substr($uuid, 0, 18) . '-' . mb_substr($uuid, -16);
+    $uuid = mb_substr($uuid, 0, 23) . '-' . mb_substr($uuid, -12);
+
+    return $uuid;
+}
+
 Php7Handler::db_connect();
 
 $db = lib\Database\DataBaseSingleton::Instance();
